Bug fixes, calculation changes
Made damage calculations more in line with the official games which use fixed point math using 4096 scale, rather than floating point math
Fixed type effectiveness bugs where it wouldn't display "doesn't affect" message and gives incorrect values
Changed how Substitute damage message is outputted. Will only say either "sub took damage" or "sub faded".
No more double types, no more needless static_cast<int> and static_cast<double> since all calculations are either floored through int truncation or rounded using half divisor addition.

+++ src/battle/BattleContext.h
@@ -33,7 +33,7 @@ struct BattleContext
 
     BattleStateFlags flags;
 
-    double damageTaken{};
+    int damageTaken{};
 
     Player* playerOne;
     Player* playerTwo;
